+++ .builders/scripts/build_wheels.py
@@ -4,7 +4,6 @@ import argparse
 import os
 import subprocess
 import sys
-from hashlib import sha256
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
@@ -26,10 +25,10 @@ if sys.platform == 'win32':
 else:
     import shlex
 
-    PY3_PATH = Path('/py3/bin/python')
-    PY2_PATH = Path('/py2/bin/python')
-    MOUNT_DIR = Path('/home')
-    ENV_FILE = Path('/.env')
+    PY3_PATH = Path(os.environ.get('DD_PY3_BUILDENV_PATH', '/py3/bin/python'))
+    PY2_PATH = Path(os.environ.get('DD_PY2_BUILDENV_PATH', '/py2/bin/python'))
+    MOUNT_DIR = Path(os.environ.get('DD_MOUNT_DIR', '/home'))
+    ENV_FILE = Path(os.environ.get('DD_ENV_FILE', '/.env'))
 
     def join_command_args(args: list[str]) -> str:
         return shlex.join(args)
@@ -66,6 +65,8 @@ def main():
         abort(f'Invalid python version: {python_version}')
 
     wheels_dir = MOUNT_DIR / 'wheels'
+    built_wheels_dir = wheels_dir / 'built'
+    external_wheels_dir = wheels_dir / 'external'
 
     # Install build dependencies
     check_process([str(python_path), '-m', 'pip', 'install', '-r', str(MOUNT_DIR / 'build_dependencies.txt')])
@@ -113,22 +114,22 @@ def main():
         check_process([
             sys.executable, '-u', str(MOUNT_DIR / 'scripts' / 'repair_wheels.py'),
             '--source-dir', str(staged_wheel_dir),
-            '--output-dir', str(wheels_dir),
+            '--built-dir', str(built_wheels_dir),
+            '--external-dir', str(external_wheels_dir),
         ])
 
     dependencies: dict[str, tuple[str, str]] = {}
-    for entry in wheels_dir.iterdir():
-        project_metadata = extract_metadata(entry)
-        project_name = normalize_project_name(project_metadata['Name'])
-        project_version = project_metadata['Version']
-        wheel_hash = sha256(entry.read_bytes()).hexdigest()
-        dependencies[project_name] = (project_version, wheel_hash)
+    for wheel_dir in wheels_dir.iterdir():
+        for entry in wheel_dir.iterdir():
+            project_metadata = extract_metadata(entry)
+            project_name = normalize_project_name(project_metadata['Name'])
+            project_version = project_metadata['Version']
+            dependencies[project_name] = project_version
 
     final_requirements = MOUNT_DIR / 'frozen.txt'
     with final_requirements.open('w', encoding='utf-8') as f:
-        for project_name, (project_version, wheel_hash) in sorted(dependencies.items()):
-            f.write(f'{project_name}=={project_version} \\\n')
-            f.write(f'  --hash=sha256:{wheel_hash}\n')
+        for project_name, project_version in sorted(dependencies.items()):
+            f.write(f'{project_name}=={project_version}\n')
 
 
 if __name__ == '__main__':